requirements.txt
streamlit>=1.31.0
selenium>=4.16.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
html5lib>=1.1
python-dotenv>=1.0.0
requests>=2.31.0